pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "leisure-recommender"
version = "0.1.0"
description = "Minimal monolith backend for personalized leisure place recommendations with UGC + moderation + local LLM integration."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "Example" }]

dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "SQLAlchemy>=2.0",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "python-multipart>=0.0.9",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "httpx>=0.27",
  "email-validator>=2.0",
]


[tool.setuptools]
package-dir = {"" = ""}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
